move to ee + remove comments
